+++ src/Table/index.js
@@ -1,0 +1,4 @@
+import { Table, Head, Body, Row, Cell } from './Table';
+
+export default Table;
+export { Table, Head, Body, Row, Cell };